Move display internals into _internal
