+++ lib/ansible/utils/display.py
@@ -18,7 +18,6 @@
 from __future__ import annotations
 
 import contextlib
-import dataclasses
 
 try:
     import curses
@@ -52,8 +51,8 @@ from ansible import constants as C
 from ansible.constants import config
 from ansible.errors import AnsibleAssertionError, AnsiblePromptInterrupt, AnsiblePromptNoninteractive, AnsibleError
 from ansible._internal._errors import _error_utils, _error_factory
-from ansible._internal import _event_formatting
-from ansible.module_utils._internal import _ambient_context, _deprecator, _messages, _no_six
+from ansible._internal import _display_utils
+from ansible.module_utils._internal import _deprecator, _messages, _no_six
 from ansible.module_utils.common.text.converters import to_bytes, to_text
 from ansible.module_utils.datatag import deprecator_from_collection_name
 from ansible._internal._datatag._tags import TrustedAsTemplate
@@ -95,6 +94,17 @@ def _is_controller_traceback_enabled(event: _traceback.TracebackEvent) -> bool:
 
 
 _traceback._is_traceback_enabled = _is_controller_traceback_enabled
+
+
+def _deprecation_warnings_enabled() -> bool:
+    """Return True if deprecation warnings are enabled for the current calling context, otherwise False."""
+    # DTFIX-FUTURE: move this capability into config using an AmbientContext-derived TaskContext (once it exists)
+    if warning_ctx := _display_utils.DeferredWarningContext.current(optional=True):
+        variables = warning_ctx._variables
+    else:
+        variables = None
+
+    return C.config.get_config_value('DEPRECATION_WARNINGS', variables=variables)
 
 
 def get_text_width(text: str) -> int:
@@ -579,7 +589,7 @@ class Display(metaclass=Singleton):
             version="2.23",
         )
 
-        msg = self._get_deprecation_message_with_plugin_info(
+        msg = _display_utils.get_deprecation_message_with_plugin_info(
             msg=msg,
             version=version,
             removed=removed,
@@ -593,70 +603,6 @@ class Display(metaclass=Singleton):
             msg = f'[DEPRECATION WARNING]: {msg}'
 
         return msg
-
-    def _get_deprecation_message_with_plugin_info(
-        self,
-        *,
-        msg: str,
-        version: str | None,
-        removed: bool = False,
-        date: str | None,
-        deprecator: _messages.PluginInfo | None,
-    ) -> str:
-        """Internal use only. Return a deprecation message and help text for display."""
-        # DTFIX-FUTURE: the logic for omitting date/version doesn't apply to the payload, so it shows up in vars in some cases when it should not
-
-        if removed:
-            removal_fragment = 'This feature was removed'
-        else:
-            removal_fragment = 'This feature will be removed'
-
-        if not deprecator or not deprecator.type:
-            # indeterminate has no resolved_name or type
-            # collections have a resolved_name but no type
-            collection = deprecator.resolved_name if deprecator else None
-            plugin_fragment = ''
-        elif deprecator.resolved_name == 'ansible.builtin':
-            # core deprecations from base classes (the API) have no plugin name, only 'ansible.builtin'
-            plugin_type_name = str(deprecator.type) if deprecator.type is _messages.PluginType.MODULE else f'{deprecator.type} plugin'
-
-            collection = deprecator.resolved_name
-            plugin_fragment = f'the {plugin_type_name} API'
-        else:
-            parts = deprecator.resolved_name.split('.')
-            plugin_name = parts[-1]
-            plugin_type_name = str(deprecator.type) if deprecator.type is _messages.PluginType.MODULE else f'{deprecator.type} plugin'
-
-            collection = '.'.join(parts[:2]) if len(parts) > 2 else None
-            plugin_fragment = f'{plugin_type_name} {plugin_name!r}'
-
-        if collection and plugin_fragment:
-            plugin_fragment += ' in'
-
-        if collection == 'ansible.builtin':
-            collection_fragment = 'ansible-core'
-        elif collection:
-            collection_fragment = f'collection {collection!r}'
-        else:
-            collection_fragment = ''
-
-        if not collection:
-            when_fragment = 'in the future' if not removed else ''
-        elif date:
-            when_fragment = f'in a release after {date}'
-        elif version:
-            when_fragment = f'version {version}'
-        else:
-            when_fragment = 'in a future release' if not removed else ''
-
-        if plugin_fragment or collection_fragment:
-            from_fragment = 'from'
-        else:
-            from_fragment = ''
-
-        deprecation_msg = ' '.join(f for f in [removal_fragment, from_fragment, plugin_fragment, collection_fragment, when_fragment] if f) + '.'
-
-        return _join_sentences(msg, deprecation_msg)
 
     @staticmethod
     def _deduplicate(msg: str, messages: set[str]) -> bool:
@@ -726,7 +672,7 @@ class Display(metaclass=Singleton):
         _skip_stackwalk = True
 
         if removed:
-            formatted_msg = self._get_deprecation_message_with_plugin_info(
+            formatted_msg = _display_utils.get_deprecation_message_with_plugin_info(
                 msg=msg,
                 version=version,
                 removed=removed,
@@ -753,7 +699,7 @@ class Display(metaclass=Singleton):
             deprecator=deprecator,
         )
 
-        if warning_ctx := _DeferredWarningContext.current(optional=True):
+        if warning_ctx := _display_utils.DeferredWarningContext.current(optional=True):
             warning_ctx.capture(deprecation)
             return
 
@@ -766,12 +712,12 @@ class Display(metaclass=Singleton):
         # This is the post-proxy half of the `deprecated` implementation.
         # Any logic that must occur in the primary controller process needs to be implemented here.
 
-        if not _DeferredWarningContext.deprecation_warnings_enabled():
+        if not _deprecation_warnings_enabled():
             return
 
         self.warning('Deprecation warnings can be disabled by setting `deprecation_warnings=False` in ansible.cfg.')
 
-        msg = _format_message(warning, _traceback.is_traceback_enabled(_traceback.TracebackEvent.DEPRECATED))
+        msg = _display_utils.format_message(warning, _traceback.is_traceback_enabled(_traceback.TracebackEvent.DEPRECATED))
         msg = f'[DEPRECATION WARNING]: {msg}'
 
         if self._deduplicate(msg, self._deprecations):
@@ -809,7 +755,7 @@ class Display(metaclass=Singleton):
             ),
         )
 
-        if warning_ctx := _DeferredWarningContext.current(optional=True):
+        if warning_ctx := _display_utils.DeferredWarningContext.current(optional=True):
             warning_ctx.capture(warning)
             return
 
@@ -822,7 +768,7 @@ class Display(metaclass=Singleton):
         # This is the post-proxy half of the `warning` implementation.
         # Any logic that must occur in the primary controller process needs to be implemented here.
 
-        msg = _format_message(warning, _traceback.is_traceback_enabled(_traceback.TracebackEvent.WARNING))
+        msg = _display_utils.format_message(warning, _traceback.is_traceback_enabled(_traceback.TracebackEvent.WARNING))
         msg = f"[WARNING]: {msg}"
 
         if self._deduplicate(msg, self._warns):
@@ -912,7 +858,7 @@ class Display(metaclass=Singleton):
             event=event,
         )
 
-        if warning_ctx := _DeferredWarningContext.current(optional=True):
+        if warning_ctx := _display_utils.DeferredWarningContext.current(optional=True):
             warning_ctx.capture(warning)
             return
 
@@ -949,7 +895,7 @@ class Display(metaclass=Singleton):
         # This is the post-proxy half of the `error` implementation.
         # Any logic that must occur in the primary controller process needs to be implemented here.
 
-        msg = _format_message(error, _traceback.is_traceback_enabled(_traceback.TracebackEvent.ERROR))
+        msg = _display_utils.format_message(error, _traceback.is_traceback_enabled(_traceback.TracebackEvent.ERROR))
         msg = f'[ERROR]: {msg}'
 
         if self._deduplicate(msg, self._errors):
@@ -1168,92 +1114,6 @@ class Display(metaclass=Singleton):
 
 
 _display = Display()
-
-
-class _DeferredWarningContext(_ambient_context.AmbientContextBase):
-    """
-    Calls to `Display.warning()` and `Display.deprecated()` within this context will cause the resulting warnings to be captured and not displayed.
-    The intended use is for task-initiated warnings to be recorded with the task result, which makes them visible to registered results, callbacks, etc.
-    The active display callback is responsible for communicating any warnings to the user.
-    """
-
-    # DTFIX-FUTURE: once we start implementing nested scoped contexts for our own bookkeeping, this should be an interface facade that forwards to the nearest
-    #               context that actually implements the warnings collection capability
-
-    def __init__(self, *, variables: dict[str, object]) -> None:
-        self._variables = variables  # DTFIX-FUTURE: move this to an AmbientContext-derived TaskContext (once it exists)
-        self._deprecation_warnings: list[_messages.DeprecationSummary] = []
-        self._warnings: list[_messages.WarningSummary] = []
-        self._seen: set[_messages.WarningSummary] = set()
-
-    @classmethod
-    def deprecation_warnings_enabled(cls) -> bool:
-        """Return True if deprecation warnings are enabled for the current calling context, otherwise False."""
-        # DTFIX-FUTURE: move this capability into config using an AmbientContext-derived TaskContext (once it exists)
-        if warning_ctx := cls.current(optional=True):
-            variables = warning_ctx._variables
-        else:
-            variables = None
-
-        return C.config.get_config_value('DEPRECATION_WARNINGS', variables=variables)
-
-    def capture(self, warning: _messages.WarningSummary) -> None:
-        """Add the warning/deprecation to the context if it has not already been seen by this context."""
-        if warning in self._seen:
-            return
-
-        self._seen.add(warning)
-
-        if isinstance(warning, _messages.DeprecationSummary):
-            self._deprecation_warnings.append(warning)
-        else:
-            self._warnings.append(warning)
-
-    def get_warnings(self) -> list[_messages.WarningSummary]:
-        """Return a list of the captured non-deprecation warnings."""
-        # DTFIX-FUTURE: return a read-only list proxy instead
-        return self._warnings
-
-    def get_deprecation_warnings(self) -> list[_messages.DeprecationSummary]:
-        """Return a list of the captured deprecation warnings."""
-        # DTFIX-FUTURE: return a read-only list proxy instead
-        return self._deprecation_warnings
-
-
-def _join_sentences(first: str | None, second: str | None) -> str:
-    """Join two sentences together."""
-    first = (first or '').strip()
-    second = (second or '').strip()
-
-    if first and first[-1] not in ('!', '?', '.'):
-        first += '.'
-
-    if second and second[-1] not in ('!', '?', '.'):
-        second += '.'
-
-    if first and not second:
-        return first
-
-    if not first and second:
-        return second
-
-    return ' '.join((first, second))
-
-
-def _format_message(summary: _messages.SummaryBase, include_traceback: bool) -> str:
-    if isinstance(summary, _messages.DeprecationSummary):
-        deprecation_message = _display._get_deprecation_message_with_plugin_info(
-            msg=summary.event.msg,
-            version=summary.version,
-            date=summary.date,
-            deprecator=summary.deprecator,
-        )
-
-        event = dataclasses.replace(summary.event, msg=deprecation_message)
-    else:
-        event = summary.event
-
-    return _event_formatting.format_event(event, include_traceback)
 
 
 def _report_config_warnings(deprecator: _messages.PluginInfo) -> None:
